+++ lib/class_core.php
@@ -71,12 +71,12 @@ class WPP_Core {
     /**
      * Maybe register site on UD
      */
-    add_action( 'wpp_post_init', array( $this, 'maybe_register_site' ) );
+    add_action( 'admin_init', array( $this, 'maybe_register_site' ), 100 );
 
     /**
      * Updated remote settings to UD
      */
-    add_action( 'wpp::save_settings', array( $this, 'update_site_settings' ) );
+    // add_action( 'wpp::save_settings', array( $this, 'update_site_settings' ) );
 
     add_filter( 'wpp_get_properties_query', array( $this, 'fix_tax_property_query' ));
 
@@ -172,69 +172,86 @@ class WPP_Core {
   }
 
   /**
-   * Register Site on UD if needed
+   * Register Site on UD, if needed. Runs on admin_init.
    *
+   *
+   * wp option delete ud_site_secret_token; wp option delete ud_site_id; wp option delete ud_site_public_key; wp transient delete wpp_state;
+   *
+   * @author potanin@UD
    */
   public function maybe_register_site() {
     global $wpdb, $wp_properties;
 
-    $table_prefix = $wpdb->prefix;
-
-    // Token set, do not attempt registration unless it cleared.
-    if( get_site_option( 'ud_site_secret_token' ) ) {
+    // Do nothing on Ajax, XMLRPC or wp-cli requests.
+    if ( ( defined( 'DOING_AJAX' ) && DOING_AJAX ) || ( defined( 'XMLRPC_REQUEST' ) && XMLRPC_REQUEST ) || ( defined( 'WP_CLI' ) && WP_CLI ) ) {
       return;
     }
 
+    // Tokens set, do not attempt registration unless it cleared.
+    if( get_site_option( 'ud_site_secret_token' ) && get_site_option( 'ud_site_id' ) && get_site_option( 'ud_site_public_key' ) ) {
+      return;
+    }
+
+    $_wpp_state = get_transient('wpp_state');
+
+    if( $_wpp_state && is_array( $_wpp_state ) && $_wpp_state[ 'registration-backoff' ] ) {
+      return;
+    }
+
+    // set registration back-off to avoid this being ran multiple times.
+    set_transient('wpp_state', array('registration-backoff'=> true ), 3600 );
+
     // Generate new secret token.
     add_site_option( 'ud_site_secret_token', $ud_site_secret_token = md5( wp_generate_password( 20 ) ) );
-    if(is_multisite()){
-      $site_url = network_site_url();
-    } else{
-      $site_url = get_site_url();
-    }
+
+    // get site ID and public_key just in case.
     $ud_site_id = get_site_option( 'ud_site_id' );
+    $ud_site_public_key = get_site_option( 'ud_site_public_key' );
 
-    $url = 'https://api.usabilitydynamics.com/product/v1/site/register';
-    $find = array( 'http://', 'https://' );
-    $replace = '';
-    $output = str_replace( $find, $replace, $site_url );
-
-    $configuration = array(
-      'searchable_attributes' => !empty( $wp_properties[ 'searchable_attributes' ] ) ? $wp_properties[ 'searchable_attributes' ] : array(),
-      'property_stats' => !empty( $wp_properties[ 'property_stats' ] ) ? $wp_properties[ 'property_stats' ] : array(),
-      'property_types' => !empty( $wp_properties[ 'property_types' ] ) ? $wp_properties[ 'property_types' ] : array(),
-      'geo_type_attributes' => !empty( $wp_properties[ 'geo_type_attributes' ] ) ? $wp_properties[ 'geo_type_attributes' ] : array(),
-      'predefined_values' => !empty( $wp_properties[ 'predefined_values' ] ) ? $wp_properties[ 'predefined_values' ] : array(),
-      'searchable_attr_fields' => !empty( $wp_properties[ 'searchable_attr_fields' ] ) ? $wp_properties[ 'searchable_attr_fields' ] : array(),
-      'admin_attr_fields' => !empty( $wp_properties[ 'admin_attr_fields' ] ) ? $wp_properties[ 'admin_attr_fields' ] : array(),
-      'numeric_attributes' => !empty( $wp_properties[ 'numeric_attributes' ] ) ? $wp_properties[ 'numeric_attributes' ] : array(),
-      'currency_attributes' => !empty( $wp_properties[ 'currency_attributes' ] ) ? $wp_properties[ 'currency_attributes' ] : array()
-    );
-
-    $data_settings = apply_filters( 'wpp::backup::data', array( 'wpp_settings' => $configuration ) );
-    $data_settings_json = json_encode( $data_settings );
+    if( defined( 'WPP_API_REGISTER_URL' ) && WPP_API_REGISTER_URL ) {
+      $_api_url = WPP_API_REGISTER_URL;
+    } else {
+      $_api_url = 'https://api.usabilitydynamics.com/product/property/site/register/v1';
+    }
 
     $args = array(
       'method' => 'POST',
-      'timeout' => 10,
+      'timeout' => 5,
       'redirection' => 5,
-      'httpversion' => '1.0',
-      'headers' => array(),
+      // 'headers' => array(),
       'body' => array(
-        'host' => $output,
+        'timestamp' => time(),
+        'host' => str_replace( array( 'http://', 'https://' ), '', is_multisite() ? network_site_url() : get_site_url() ),
         'ud_site_secret_token' => $ud_site_secret_token,
+        'ud_site_public_key' => $ud_site_public_key,
         'ud_site_id' => ( $ud_site_id ? $ud_site_id : '' ),
-        'db_name' => defined( 'DB_NAME' ) ? DB_NAME : null,
-        'home_url' => $site_url,
+        'db_hash' => md5( defined( 'DB_NAME' ) ? DB_NAME : null ) . '-' . md5( isset( $wpdb->prefix ) ? $wpdb->prefix : null),
+        'deployment_hash' => md5( is_multisite() ? network_site_url() : get_site_url() ) . '-' . md5( defined( 'DB_NAME' ) ? DB_NAME : null ) . '-' . md5( isset( $wpdb->prefix ) ? $wpdb->prefix : null),
+        'home_url' => is_multisite() ? network_site_url() : get_site_url(),
         'xmlrpc_url' => site_url( '/xmlrpc.php' ),
-        'table_refix' => isset( $table_prefix ) ? $table_prefix : null,
-        'wpp_settings' => $data_settings_json,
         'user_id' => get_current_user_id(),
-        'message' => "Hello, I'm WP-property plugin. Give me ID, please."
+        'message' => "Hello, I'm WP-Property plugin. Give me [ud_site_id] and [ud_site_public_key], please.",
+
+        'wpp_settings' => json_encode( apply_filters( 'wpp::backup::data', array( 'wpp_settings' => array(
+          'searchable_attributes' => !empty( $wp_properties[ 'searchable_attributes' ] ) ? $wp_properties[ 'searchable_attributes' ] : array(),
+          'property_stats' => !empty( $wp_properties[ 'property_stats' ] ) ? $wp_properties[ 'property_stats' ] : array(),
+          'property_types' => !empty( $wp_properties[ 'property_types' ] ) ? $wp_properties[ 'property_types' ] : array(),
+          'geo_type_attributes' => !empty( $wp_properties[ 'geo_type_attributes' ] ) ? $wp_properties[ 'geo_type_attributes' ] : array(),
+          'predefined_values' => !empty( $wp_properties[ 'predefined_values' ] ) ? $wp_properties[ 'predefined_values' ] : array(),
+          'searchable_attr_fields' => !empty( $wp_properties[ 'searchable_attr_fields' ] ) ? $wp_properties[ 'searchable_attr_fields' ] : array(),
+          'admin_attr_fields' => !empty( $wp_properties[ 'admin_attr_fields' ] ) ? $wp_properties[ 'admin_attr_fields' ] : array(),
+          'numeric_attributes' => !empty( $wp_properties[ 'numeric_attributes' ] ) ? $wp_properties[ 'numeric_attributes' ] : array(),
+          'currency_attributes' => !empty( $wp_properties[ 'currency_attributes' ] ) ? $wp_properties[ 'currency_attributes' ] : array()
+        ) ) ) ),
+
+        // @todo Depreciate these - potanin@UD
+        'db_name' => defined( 'DB_NAME' ) ? DB_NAME : null,
+        'table_refix' => isset( $wpdb->prefix ) ? $wpdb->prefix : null
+
       ),
     );
 
-    $response = wp_remote_post( $url, $args );
+    $response = wp_remote_post( $_api_url, $args );
 
     if( wp_remote_retrieve_response_code( $response ) === 200 && !is_wp_error( $response ) ) {
 
@@ -258,11 +275,11 @@ class WPP_Core {
 
   /**
    * Update remote settings
+   *
    */
   public function update_site_settings() {
     global $wpdb, $wp_properties;
 
-    $table_prefix = $wpdb->prefix;
     $ud_site_secret_token = get_site_option('ud_site_secret_token');
     if(is_multisite()){
       $site_url = network_site_url();
@@ -301,10 +318,10 @@ class WPP_Core {
             'host' => $output,
             'ud_site_secret_token' => $ud_site_secret_token,
             'ud_site_id' => $ud_site_id,
-	    'ud_site_public_key' => $ud_site_public_key,
+	          'ud_site_public_key' => $ud_site_public_key,
             'db_name' => DB_NAME,
             'home_url' => $site_url,
-            'table_refix' => $table_prefix,
+            'table_refix' => $wpdb->prefix,
             'wpp_settings' => $data_settings_json,
             'message' => "Hello, I'm WP-property plugin. Give me ID, please."
         ),